pyproject.toml
[tool.poetry]
name = "second-brain-agent"
version = "0.1.0"
description = "Second Brain AI Agent"
authors = ["Frederic Lepied <flepied@gmail.com>"]
license = "GPL 3.0"
readme = "README.md"
repository = "https://www.github.com/flepied/second-brain-agent"

[tool.poetry.group.dev.dependencies]
pyproject-pre-commit = "^0.0.17"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]  # main dependency group
python = ">=3.8.1,<3.12"

bs4 = "*"
chromadb = "*"
huggingface_hub = "*"
langchain =  { git = "https://github.com/raymond-yuan/langchain.git", branch = "ryuan/relevance_refactor" }
openai = "*"
pymupdf = "*"
python-dotenv = "*"
sentence-transformers = "*"
streamlit = "*"
tiktoken = "*"
unstructured = "*"
youtube-transcript-api = "*"

[tool.poetry.group.test]  # This part can be left out

[tool.poetry.group.test.dependencies]
pytest = "*"
pylint = "*"
